backend/requirements.txt
fastapi
uvicorn[standard]
gunicorn
websockets
sqlalchemy
psycopg2-binary
python-dotenv
python-multipart
python-jose[cryptography]
passlib[bcrypt]